+++ app/src/main/java/org/akvo/caddisfly/ui/SwatchActivity.java
@@ -38,5 +38,9 @@ public class SwatchActivity extends AppCompatActivity {
                     .commit();
 
         }
+
+        if (getSupportActionBar() != null) {
+            getSupportActionBar().setDisplayUseLogoEnabled(false);
+        }
     }
 }