+++ app/src/main/java/org/akvo/caddisfly/ui/VerboseResultFragment.java
@@ -72,6 +72,7 @@ public class VerboseResultFragment extends DialogFragment {
         TextView swatchRgbValue = (TextView) view.findViewById(R.id.swatchRgbValue);
         TextView textDimension = (TextView) view.findViewById(R.id.textDimension);
         TextView textDistance = (TextView) view.findViewById(R.id.textDistance);
+        TextView textQuality = (TextView) view.findViewById(R.id.textQuality);
 
         imageExtract.setImageBitmap(mExtractBitmap);
         imagePhoto.setImageBitmap(mPhotoBitmap);
@@ -87,18 +88,23 @@ public class VerboseResultFragment extends DialogFragment {
         Bundle resultValue = ColorUtils.getPpmValue(mExtractBitmap, mainApp.currentTestInfo, Config.SAMPLE_CROP_LENGTH_DEFAULT);
         double result = resultValue.getDouble(Config.RESULT_VALUE_KEY, -1);
         int color = resultValue.getInt(Config.RESULT_COLOR_KEY, 0);
-        int swatchColor = resultValue.getInt("matchedColor", 0);
+        int swatchColor = resultValue.getInt("MatchedColor", 0);
+
+
+        textQuality.setText(String.format("quality: %.0f%%", resultValue.getDouble("quality", 0)));
 
         if (result > -1) {
             textResult.setText(String.format("%s : %.2f %s", mainApp.currentTestInfo.getName("en"),
                     result, mainApp.currentTestInfo.getUnit()));
             textDistance.setText(String.format("distance: %.2f", resultValue.getDouble("Distance")));
             buttonSwatchColor.setBackgroundColor(resultValue.getInt("MatchedColor", 0));
+            swatchRgbValue.setText(String.format("r: %s", ColorUtils.getColorRgbString(swatchColor)));
+        } else {
+            textResult.setText(String.format("%s", mainApp.currentTestInfo.getName("en")));
         }
         buttonColorExtract.setBackgroundColor(color);
 
-        textColorRgb.setText(String.format("rgb: %s", ColorUtils.getColorRgbString(color)));
-        swatchRgbValue.setText(String.format("rgb: %s", ColorUtils.getColorRgbString(swatchColor)));
+        textColorRgb.setText(String.format("r: %s", ColorUtils.getColorRgbString(color)));
 
         return view;
     }